+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## Fixed
+- Fix "state" field having values of both "Canceled" and "Cancelled" in BillingAgreement model
 
 ## [0.13.0] - 2019-11-11
 ## Added